Prevent sleeping in atomic context
The function rtl_ips_nic_on() calls cancel_delayed_work_sync()
which sleeps to wait for the workqueue to flush. The kernel
generates a warning when a sleep occurs in an atomic context.

This now selects between the synchronous and asyncrhonous cancel
operations based on whether or not we are atomic.

Bug: 413